bq_firebase/pyproject.toml
[project]
name = "firebase-analytics-demo"
version = "0.1.0"
description = "Firebase Analytics BigQuery Demo with 10 Progressive Examples"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.25.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "db-dtypes>=1.2.0",
    "flask>=3.0.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.116.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = []
